requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alibabacloud-credentials==1.0.1
alibabacloud-credentials-api==1.0.0
alibabacloud-dm20151123==1.2.3
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.3
alibabacloud-openapi-util==0.2.2
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.3.14
alibabacloud-tea-util==0.3.13
alibabacloud-tea-xml==0.0.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
